requirements.txt
numpy
Pillow
pytorch
matplotlib
urllib
os
zipfile
opencv-python
moviepy
torch
torchvision
